Create aaaa.java
/*
Problem Statement of this Pucks Vending Machine:

There are 30 hockey teams in the NHL (National Hockey League). Some stores have vending machines 
that dispense miniature team hockey pucks for a toonie ($2) each. When you put in a toonie you never 
know which puck you will get; any one of the 30 team pucks is as likely as any other to be dispensed by the machine. 

They are given out at random. 

For this exercise we will limit the teams to 10.   
 
Your job is to write a program to simulate the dispensing of NHL miniature pucks 
until one of each 10 miniature team hockey pucks is dispensed. 
 
Your program should proceed as follows: 

1.	Display a welcome message and ask the user to enter their name.   
2.	Store your 10 favorite hockey teams’ name in an array of String. Assign the team names in the declaration statement directly. 
3.	Your program should loop(use a while loop) until at least one miniature puck of each team has been dispensed. Create an integer 
        array of size 10, which will serve as a counter array to keep track of the number of each team puck dispensed by the vending 
        machine. You will need to use the Math.random()function to randomly dispense a miniature puck. The Math.random() method 
        returns a double value with a positive sign, greater than or equal to 0.0 and less than 1.0.  
4.	Once you have accumulated at least one of each puck, display how many of each team puck you had to purchase, the total number 
        of pucks purchased and the total cost in a personalized message.  


 * 
 */